refactor(editor): keep popout action out of EditorPanel
