+++ src/renderer/src/components/editor/EditorPanelShell.tsx
@@ -11,6 +11,7 @@ import type { EditorToggleValue } from './EditorViewToggle'
 import { getUntitledFileRoot } from './untitled-file-rename-path'
 import { translate } from '@/i18n/i18n'
 import type { ArtifactWriteRequest } from '../../../../shared/artifacts'
+import { createEditorPopoutAction } from '@/components/editor-popout/editor-popout-action'
 
 type EditorPanelRenderModel = ReturnType<typeof getEditorPanelRenderModel>
 
@@ -43,7 +44,6 @@ type EditorPanelShellProps = {
   onToggleMarkdownFrontmatter: () => void
   onExportMarkdownToPdf: () => void
   createMarkdownArtifactRequest?: () => Promise<ArtifactWriteRequest>
-  onOpenMarkdownInNewWindow?: () => void
   onContentChange: (content: string) => void
   onContentChangeForFile: (file: OpenFile, content: string) => void
   onDirtyStateHint: (dirty: boolean) => void
@@ -85,7 +85,6 @@ export function EditorPanelShell({
   onToggleMarkdownFrontmatter,
   onExportMarkdownToPdf,
   createMarkdownArtifactRequest,
-  onOpenMarkdownInNewWindow,
   onContentChange,
   onContentChangeForFile,
   onDirtyStateHint,
@@ -97,6 +96,15 @@ export function EditorPanelShell({
   onRenameConfirm,
   markdownAnnotationsEnabled
 }: EditorPanelShellProps): JSX.Element {
+  const openMarkdownInNewWindow = createEditorPopoutAction({
+    getState: useAppStore.getState,
+    file: activeFile,
+    fileContent: fileContents[activeFile.id],
+    content: editorDrafts[activeFile.id],
+    viewMode: model.mdViewMode,
+    showFrontmatter: markdownFrontmatterVisible
+  })
+
   return (
     <div ref={panelRef} className="flex flex-col flex-1 min-w-0 min-h-0">
       {!model.isCombinedDiff && activeFile.mode !== 'check-details' && (
@@ -133,7 +141,7 @@ export function EditorPanelShell({
           onToggleMarkdownFrontmatter={onToggleMarkdownFrontmatter}
           onExportMarkdownToPdf={onExportMarkdownToPdf}
           createMarkdownArtifactRequest={createMarkdownArtifactRequest}
-          onOpenMarkdownInNewWindow={onOpenMarkdownInNewWindow}
+          onOpenMarkdownInNewWindow={openMarkdownInNewWindow}
         />
       )}
       <Suspense fallback={<EditorLoadingFallback />}>
